Ball deflection and trail, improved movement parameters
Enabled paddles to force ball to move left or right depending on collision position with paddle. Added trail renderer for the ball. Added more control parameters to optimize the game. Updated scenes and prefabs to include newest changes.

+++ Assets/Scripts/PlayerPaddleMover.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class PlayerPaddleMover : MonoBehaviour
 {  
     [Tooltip("How far to the left/right can the paddle move?")]
-    public int HorizontalPositionThreshold = 4;
+    public float HorizontalPositionThreshold = 4;
 
     [Tooltip("Horizontal force to act upon the paddle (in Joystick or LeftRight mode).")]
     public int SpeedForce = 500;
